* Reimplementação de mais alguns eventos, diminuindo o número final de arquivos compilados. * Remoção de componentes (painéis) inúteis. * Correção da regex que verifica o "sim/não" das condições. * Correção de erro no socket que verificava comandos de controle de música. * Adição de testes de integridade.

+++ trunk/src/com/playmusica/PlayMusicaBD.java
@@ -139,6 +139,9 @@ public class PlayMusicaBD {
      * @return int Contendo o numero de linhas afetadas.
      */
     public static int excluirPelaBk(PlayMusica playMusica, Transacao t) throws Exception {
+        if (playMusica.getPlaylist() == null) {
+            throw new Exception("PlayList não carregada.");
+        }
         SQL sql = new SQL();
         sql.add("DELETE FROM " + TBL);
         sql.add("WHERE playlist = :playlist  ");
